refactor: Organize imports, remove unused code, and enhance service management interfaces
