+++ velithon/vsp/__init__.py
@@ -1,6 +1,7 @@
 from .client import VSPClient
 from .manager import VSPManager, WorkerType
-from .mesh import ServiceInfo, ServiceMesh
+from .mesh import ServiceMesh
+from .service import ServiceInfo
 from .message import VSPMessage
 from .protocol import VSPProtocol
